// contracts/container-runtime/v1/generated/container-runtime.protobuf.ts
export const containerRuntimeProtobufDescriptor: INamespace = {
  "nested": {
    "cinatoken": {
      "nested": {
        "container": {
          "nested": {
            "runtime": {
              "nested": {
                "v1": {
                  "nested": {
// ...
                    "OperationEnvelope": {
                      "fields": {
// ...
                        "input": {
                          "id": 9,
                          "options": {
                            "json_name": "input"
                          },
                          "type": "OperationInput"
                        },
                        "shard": {
                          "id": 10,
                          "options": {
                            "json_name": "shard"
                          },
                          "type": "OperationShard"
                        },
                        "trace_id": {
                          "id": 11,
                          "options": {
                            "json_name": "trace_id"
                          },
                          "type": "string"
                        }
                      }
                    },
// ...
                  }
                }
              }
            }
          }
        }
      }
    }
  }
};